feat(**hipcomp.cmake): + arg BUILD_STATIC
  rapids_cpm_hipcomp( [USE_PROPRIETARY_BINARY <ON|OFF>]
                     [BUILD_EXPORT_SET <export-name>]
                     [INSTALL_EXPORT_SET <export-name>]
NEW:                 [BUILD_STATIC] <build-static>
                     [<CPM_ARGS> ...])

The argument `BUILD_STATIC` can be used to specify if the libhipcomp library
should be built as a static or shared object. Defaults to `ON`. You can also
use `RAPIDS_CMAKE_HIPCOMP_BUILD_STATIC` to specify this option. If both are
specified, the argument overrules the environment variable.