Remove now superfluous fmt patch (#790)
With the migration to rapids-logger in 25.04, we no longer compile fmt by default, and in cases where we wish to do so explicitly (i.e. when we turn off the symbol controls) we still only compile it in host TUs. Therefore, we can drop this patch, which currently precludes the usage of a system fmt library if one has already been built. Non-RAPIDS consumers who need to build fmt in nvcc without warnings can apply this patch themselves with an override.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/rapidsai/rapids-cmake/pull/790