Add build patch functionality (#801)
This enables patches that only affect the build process of a project and not its exported headers or runtime functionality. If all of a project's patches are build patches, it is not required to be downloaded and can be found with `find_package()`.

Resolves https://github.com/rapidsai/rapids-cmake/issues/799

TODO:

- [x] Add tests for `rapids_cpm_find()`

Authors:
  - Kyle Edwards (https://github.com/KyleFromNVIDIA)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/rapidsai/rapids-cmake/pull/801